+++ src/snorlax/protocol/internet/version4/control/message/module.c
@@ -16,8 +16,8 @@ static void internet_control_message_protocol_version4_module_func_timestamp_mes
 static void internet_control_message_protocol_version4_module_func_information_message_debug(internet_control_message_protocol_version4_module_t * module, FILE * stream, internet_control_message_protocol_version4_context_t * context);
 
 static internet_control_message_protocol_version4_module_t * internet_control_message_protocol_version4_module_func_rem(internet_control_message_protocol_version4_module_t * module);
-static int32_t internet_control_message_protocol_version4_module_func_deserialize(internet_control_message_protocol_version4_module_t * module, protocol_packet_t * packet, uint32_t packetlen, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t ** context);
-static int32_t internet_control_message_protocol_version4_module_func_serialize(internet_control_message_protocol_version4_module_t * module, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t * context, protocol_packet_t ** packet, uint32_t * packetlen);
+static int32_t internet_control_message_protocol_version4_module_func_deserialize(internet_control_message_protocol_version4_module_t * module, protocol_packet_t * packet, uint64_t packetlen, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t ** context);
+static int32_t internet_control_message_protocol_version4_module_func_serialize(internet_control_message_protocol_version4_module_t * module, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t * context, protocol_packet_t ** packet, uint64_t * packetlen);
 static void internet_control_message_protocol_version4_module_func_debug(internet_control_message_protocol_version4_module_t * module, FILE * stream, internet_control_message_protocol_version4_context_t * context);
 
 static internet_control_message_protocol_version4_module_func_t func = {
@@ -47,7 +47,7 @@ static internet_control_message_protocol_version4_module_t * internet_control_me
     return nil;
 }
 
-static int32_t internet_control_message_protocol_version4_module_func_deserialize(internet_control_message_protocol_version4_module_t * module, protocol_packet_t * packet, uint32_t packetlen, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t ** context) {
+static int32_t internet_control_message_protocol_version4_module_func_deserialize(internet_control_message_protocol_version4_module_t * module, protocol_packet_t * packet, uint64_t packetlen, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t ** context) {
 #ifndef   RELEASE
     snorlaxdbg(module == nil, false, "criticial", "");
     snorlaxdbg(packet == nil, false, "criticial", "");
@@ -73,7 +73,7 @@ static int32_t internet_control_message_protocol_version4_module_func_deserializ
     return success;
 }
 
-static int32_t internet_control_message_protocol_version4_module_func_serialize(internet_control_message_protocol_version4_module_t * module, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t * context, protocol_packet_t ** packet, uint32_t * packetlen) {
+static int32_t internet_control_message_protocol_version4_module_func_serialize(internet_control_message_protocol_version4_module_t * module, internet_protocol_version4_context_t * parent, internet_control_message_protocol_version4_context_t * context, protocol_packet_t ** packet, uint64_t * packetlen) {
     snorlaxdbg(true, false, "implement", "");
 
     return fail;
