+++ src/snorlax/protocol/internet/version6/module/router.c
@@ -1,8 +1,8 @@
 #include "../../version6.h"
 
 typedef internet_protocol_version6_module_router_t * (*internet_protocol_version6_module_router_func_rem_t)(internet_protocol_version6_module_router_t *);
-typedef int32_t (*internet_protocol_version6_module_router_func_deserialize_t)(internet_protocol_version6_module_router_t *, protocol_packet_t *, uint32_t, protocol_context_t *, internet_protocol_version6_context_t **);
-typedef int32_t (*internet_protocol_version6_module_router_func_serialize_t)(internet_protocol_version6_module_router_t *, protocol_context_t *, internet_protocol_version6_context_t *, protocol_packet_t **, uint32_t *);
+typedef int32_t (*internet_protocol_version6_module_router_func_deserialize_t)(internet_protocol_version6_module_router_t *, protocol_packet_t *, uint64_t, protocol_context_t *, internet_protocol_version6_context_t **);
+typedef int32_t (*internet_protocol_version6_module_router_func_serialize_t)(internet_protocol_version6_module_router_t *, protocol_context_t *, internet_protocol_version6_context_t *, protocol_packet_t **, uint64_t *);
 typedef void (*internet_protocol_version6_module_router_func_debug_t)(internet_protocol_version6_module_router_t *, FILE *, internet_protocol_version6_context_t *);
 
 static internet_protocol_version6_module_router_func_t func = {
